+++ daengle-vet-api/build.gradle
@@ -2,3 +2,19 @@ bootJar {
     enabled = true
 }
 
+jar {
+    enabled = false
+}
+
+dependencies {
+    implementation(project(':daengle-domain'))
+    implementation(project(':daengle-persistence'))
+    implementation(project(':daengle-auth'))
+
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+
+    implementation 'org.springframework:spring-tx'
+
+    compileOnly 'org.projectlombok:lombok'
+    annotationProcessor 'org.projectlombok:lombok'
+}
